test for tredje case utregning
