+++ src/komponenter/kalkulator/utregninger.ts
@@ -83,193 +83,6 @@ export const sumPermitteringerOgFravær = (
     return oversikt;
 };
 
-//AGP2FUNKSJONER
-
-export enum ArbeidsgiverPeriode2Resulatet {
-    NÅDD_AGP2,
-    LØPENDE_IKKE_NÅDD_AGP2,
-    IKKE_LØPENDE_IKKE_NÅDD_AGP2,
-}
-
-export interface InformasjonOmGjenståendeDagerOgPeriodeAGP2 {
-    sluttDato: Date | undefined;
-    gjenståendePermitteringsDager: number;
-    brukteDager: number;
-    type: ArbeidsgiverPeriode2Resulatet;
-}
-
-export const finnDatoAGP2 = (
-    tidslinje: DatoMedKategori[],
-    tidligsteDatoAGP2: Date,
-    erLøpende: boolean,
-    dagensDato: Date
-): InformasjonOmGjenståendeDagerOgPeriodeAGP2 => {
-    const statusPermittering1muligAGP2 = finnOversiktOverBruktPermitteringIGittIntervall(
-        tidligsteDatoAGP2,
-        tidslinje
-    );
-    const gjenVærendePermittering = finnGjenståendePermitteringsDager(
-        statusPermittering1muligAGP2
-    );
-    if (gjenVærendePermittering <= 0) {
-        return {
-            sluttDato: tidligsteDatoAGP2,
-            brukteDager:
-                statusPermittering1muligAGP2.dagerPermittert -
-                statusPermittering1muligAGP2.dagerAnnetFravær,
-            gjenståendePermitteringsDager: 0,
-            type: 0,
-        };
-    }
-    if (erLøpende) {
-        return finnDatoAGP2LøpendPermittering(
-            tidslinje,
-            tidligsteDatoAGP2,
-            gjenVærendePermittering
-        );
-    } else {
-        const oversikt = finnPeriodeForÅBrukeGjenståendeDager(
-            tidslinje,
-            tidligsteDatoAGP2,
-            dagensDato
-        );
-        return oversikt;
-    }
-};
-
-//må legge til grensetilfeller dersom AGP2 inntreffer for langt fram i tid. Viktig å teste - setDato() her.
-export const finnDatoAGP2LøpendPermittering = (
-    tidslinje: DatoMedKategori[],
-    tidligsteDatoAGP2: Date,
-    gjenStåendeDager: number
-): InformasjonOmGjenståendeDagerOgPeriodeAGP2 => {
-    let sluttDato: Date | undefined = new Date(tidligsteDatoAGP2);
-    let dagerGjenstår = gjenStåendeDager;
-    while (dagerGjenstår > 0) {
-        sluttDato.setDate(sluttDato.getDate() + dagerGjenstår);
-        const statusPermittering1muligAGP2 = finnOversiktOverBruktPermitteringIGittIntervall(
-            sluttDato,
-            tidslinje
-        );
-        dagerGjenstår =
-            finnGjenståendePermitteringsDager(statusPermittering1muligAGP2) -
-            dagerGjenstår;
-        if (sluttDato > tidslinje[tidslinje.length - 1].dato) {
-            sluttDato = undefined;
-            break;
-        }
-    }
-    return {
-        sluttDato: sluttDato,
-        gjenståendePermitteringsDager: gjenStåendeDager,
-        brukteDager: 210 - gjenStåendeDager,
-        type: 1,
-    };
-};
-
-const finnGjenståendePermitteringsDager = (
-    oversikt: OversiktOverBrukteOgGjenværendeDager
-) => {
-    return 7 * 30 - (oversikt.dagerPermittert - oversikt.dagerAnnetFravær);
-};
-
-//regner ut differansen i gjenstående permitteringsdager i forhold til antall dager man kan plassere permitteringsdatoene på
-//eksempel hvis man har 20 gjenstående permitteringsdager man kan bruke fordelt på 5 dager, returnerer funksjonen 20-5
-export const finnOverskuddAvMuligePermitteringsdager = (
-    sisteDagIAktuellPeriode: Date,
-    tidslinje: DatoMedKategori[],
-    oversiktOverBrukteDagerIPeriode: OversiktOverBrukteOgGjenværendeDager,
-    dagensDato: Date
-) => {
-    const ubrukteDagerIPeriode = finnGjenståendePermitteringsDager(
-        oversiktOverBrukteDagerIPeriode
-    );
-    const dagerMellomDagensDatoOgSisteDagIAktuellPeriode = antalldagerGått(
-        dagensDato,
-        sisteDagIAktuellPeriode
-    );
-    return (
-        ubrukteDagerIPeriode - dagerMellomDagensDatoOgSisteDagIAktuellPeriode
-    );
-};
-
-export const finnPeriodeForÅBrukeGjenståendeDager = (
-    tidslinje: DatoMedKategori[],
-    tidligsteDatoAGP2: Date,
-    dagensDato: Date
-): InformasjonOmGjenståendeDagerOgPeriodeAGP2 => {
-    const førstePermitteringInnenForTidligsteAGP2Periode = tidslinje.find(
-        (datoMedKategori) =>
-            datoMedKategori.kategori === 0 &&
-            datoMedKategori.dato >= finnDato18MndTilbake(tidligsteDatoAGP2)
-    );
-    if (førstePermitteringInnenForTidligsteAGP2Periode === undefined) {
-        return {
-            sluttDato: undefined,
-            brukteDager: 0,
-            gjenståendePermitteringsDager: 210,
-            type: 0,
-        };
-    }
-    const sisteDagIAktuellPeriode = finnDato18MndFram(
-        førstePermitteringInnenForTidligsteAGP2Periode.dato
-    );
-    let oversiktOverBrukteDagerIPeriode = finnOversiktOverBruktPermitteringIGittIntervall(
-        sisteDagIAktuellPeriode,
-        tidslinje
-    );
-    let overskuddAvPermitteringsdagerItidsintervall = finnOverskuddAvMuligePermitteringsdager(
-        sisteDagIAktuellPeriode,
-        tidslinje,
-        oversiktOverBrukteDagerIPeriode,
-        dagensDato
-    );
-    while (overskuddAvPermitteringsdagerItidsintervall > 0) {
-        sisteDagIAktuellPeriode.setDate(
-            sisteDagIAktuellPeriode.getDate() +
-                overskuddAvPermitteringsdagerItidsintervall
-        );
-        let oversiktOverBrukteDagerIPeriode = finnOversiktOverBruktPermitteringIGittIntervall(
-            sisteDagIAktuellPeriode,
-            tidslinje
-        );
-        overskuddAvPermitteringsdagerItidsintervall = finnOverskuddAvMuligePermitteringsdager(
-            sisteDagIAktuellPeriode,
-            tidslinje,
-            oversiktOverBrukteDagerIPeriode,
-            dagensDato
-        );
-        if (sisteDagIAktuellPeriode > tidslinje[0].dato) {
-            break;
-        }
-    }
-    const gjeldendeSlutt =
-        sisteDagIAktuellPeriode > tidslinje[0].dato
-            ? sisteDagIAktuellPeriode
-            : undefined;
-    let sluttOversikt = gjeldendeSlutt
-        ? finnOversiktOverBruktPermitteringIGittIntervall(
-              gjeldendeSlutt,
-              tidslinje
-          ).dagerPermittert -
-          finnOversiktOverBruktPermitteringIGittIntervall(
-              gjeldendeSlutt,
-              tidslinje
-          ).dagerAnnetFravær
-        : 210;
-    return {
-        sluttDato: gjeldendeSlutt,
-        gjenståendePermitteringsDager: antalldagerGått(
-            dagensDato,
-            gjeldendeSlutt
-        ),
-        brukteDager: sluttOversikt,
-        type: 2,
-    };
-};
-
-//AGP2FUNKSJONER SLUTT
-
 export const finnOversiktOverBruktPermitteringIGittIntervall = (
     sisteDatoIAktuellPeriode: Date,
     tidslinje: DatoMedKategori[]
